% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{find_compiled_model}
\alias{find_compiled_model}
\title{Find the path to the compiled models}
\usage{
find_compiled_model()
}
\value{
a character string
}
\description{
Find the path to the compiled models
}
\examples{
find_compiled_model()
}

