% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albatross.R
\name{walb_deb_log_food}
\alias{walb_deb_log_food}
\title{Teixeira albatross model with logistic food in nestling stage}
\usage{
walb_deb_log_food(t, states, params)
}
\arguments{
\item{t}{vector of times}

\item{states}{vector of initial state values}

\item{params}{vector of parameters}
}
\value{
derivatives of the ODE
}
\description{
Teixeira albatross model with logistic food in nestling stage
}
